[Refactor] Extract stream context manager from transaction manager
