+++ be/test/runtime/stream_load/stream_load_context_test.cpp
@@ -23,8 +23,8 @@
 #include "base/utility/defer_op.h"
 #include "runtime/exec_env.h"
 #include "runtime/message_body_sink.h"
+#include "runtime/stream_load/stream_context_mgr.h"
 #include "runtime/stream_load/stream_load_context_handle.h"
-#include "runtime/stream_load/transaction_mgr.h"
 
 namespace starrocks {
