+++ build-support/exec_env_header_include_allowlist.txt
@@ -1,4 +1,3 @@
 # Headers that still require a complete ExecEnv or GlobalEnv definition.
-src/runtime/stream_load/transaction_mgr.h
 test/testutil/init_test_env.h
 test/storage/lake/test_util.h
